Break out GBPTree read and write tests to separate test class
